Fixed end_object find code
Previously the end_object for the data pack was found by searching for '</script>' and then going 4 characters back, however google in a recent update has added , 'sideChannel: {}});' to the end of the data pack, which throws it off. To fix this the end_object finding script first searches for '</script>' and then searches for the first ']' to the left of that closing script tag. This should be more flexible.