Improved exception handling
Previously if the data unpacking failed it would tell the user that the URL could not be opened. But this is the wrong exception. So i fixed this by splitting up the data un packing and url opening into seperate parts so each can have their own exception. This should make it easier to identify what has gone wrong.